M_13_DRF/mysite/requirement.txt
asgiref==3.5.2
attrs==23.1.0
Django==4.0.6
django-filter==23.2
djangorestframework==3.14.0
docutils==0.20.1
drf-spectacular==0.26.2
inflection==0.5.1
jsonschema==4.17.3
Pillow==9.5.0
pyrsistent==0.19.3
pytz==2022.7.1
PyYAML==6.0
sqlparse==0.4.2
uritemplate==4.1.1
